+++ src/components/modals/ChangeNetwork.tsx
@@ -7,6 +7,7 @@ import { IconXDAI } from '../Icons/XDAI';
 import { H4, B, neutralColors, brandColors } from '@giveth/ui-design-system';
 import { useWeb3React } from '@web3-react/core';
 import { mediaQueries } from '@/utils/constants';
+import { ETheme, useGeneral } from '@/context/general.context';
 
 interface IChangeNetworkModalProps extends IModal {
 	targetNetwork: number;
@@ -18,6 +19,7 @@ export const ChangeNetworkModal: FC<IChangeNetworkModalProps> = ({
 	targetNetwork,
 }) => {
 	const { chainId } = useWeb3React();
+	const { theme } = useGeneral();
 
 	useEffect(() => {
 		if (chainId === targetNetwork) {
@@ -36,7 +38,7 @@ export const ChangeNetworkModal: FC<IChangeNetworkModalProps> = ({
 				) : (
 					<IconXDAI size={64} />
 				)}
-				<Title>Switch to {NetworkName}</Title>
+				<Title theme={theme}>Switch to {NetworkName}</Title>
 				<Desc>Please switch your wallet network to {NetworkName}.</Desc>
 			</ChangeNetworkModalContainer>
 		</Modal>
@@ -47,20 +49,16 @@ const ChangeNetworkModalContainer = styled.div`
 	width: 500px;
 	padding: 62px 60px;
 	color: ${brandColors.giv[700]};
-	${mediaQueries['mobileS']} {
-		width: 100%;
-	}
-	${mediaQueries['mobileM']} {
-		width: 100%;
-	}
-	${mediaQueries['mobileL']} {
-		width: 100%;
+	width: 100%;
+	${mediaQueries['tablet']} {
+		width: 500px;
 	}
 `;
 
 const Title = styled(H4)`
 	margin: 18px 0 24px;
-	color: ${brandColors.giv[700]};
+	color: ${props =>
+		props.theme === ETheme.Dark ? 'white' : brandColors.giv[700]};
 `;
 
 const Desc = styled(B)``;
