+++ src/components/views/donate/TokenPicker.tsx
@@ -7,7 +7,7 @@ import Select, {
 	StylesConfig,
 } from 'react-select';
 import Image from 'next/image';
-import { useDeviceDetect } from '@/utils';
+import useDeviceDetect from '@/hooks/useDeviceDetect';
 import { defaultTheme } from 'react-select';
 import { neutralColors, P, B, brandColors } from '@giveth/ui-design-system';
 import XIcon from '/public/images/x-icon.svg';
@@ -312,19 +312,17 @@ const Control = (props: any) => {
 const Menu = (props: JSX.IntrinsicElements['div']) => {
 	const shadow = 'hsla(218, 50%, 10%, 0.1)';
 	return (
-		<>
-			<div
-				style={{
-					backgroundColor: 'white',
-					borderRadius: 4,
-					boxShadow: `0 0 0 1px ${shadow}, 0 4px 11px ${shadow}`,
-					marginTop: 8,
-					position: 'absolute',
-					zIndex: 2,
-				}}
-				{...props}
-			/>
-		</>
+		<div
+			style={{
+				backgroundColor: 'white',
+				borderRadius: 4,
+				boxShadow: `0 0 0 1px ${shadow}, 0 4px 11px ${shadow}`,
+				marginTop: 8,
+				position: 'absolute',
+				zIndex: 2,
+			}}
+			{...props}
+		/>
 	);
 };
 const Blanket = (props: JSX.IntrinsicElements['div']) => (
