+++ src/utils/index.tsx
@@ -31,41 +31,6 @@ export function pollEvery(fn: Function, delay: any) {
 	};
 }
 
-// TODO: fix this so it relates the change of width
-export function _useDeviceDetect() {
-	const [size, setSize] = React.useState([0, 0]);
-	React.useLayoutEffect(() => {
-		function updateSize() {
-			setSize([window.innerWidth, window.innerHeight]);
-		}
-		window.addEventListener('resize', updateSize);
-		updateSize();
-		return () => window.removeEventListener('resize', updateSize);
-	}, []);
-	return { isMobile: size[0] <= deviceSize.mobileL };
-}
-
-export function useDeviceDetect() {
-	const [isMobile, setMobile] = React.useState(false);
-
-	React.useEffect(() => {
-		const userAgent =
-			typeof window.navigator === 'undefined' ? '' : navigator.userAgent;
-		let mobile = Boolean(
-			userAgent.match(
-				/Android|BlackBerry|iPhone|iPod|Opera Mini|IEMobile/i,
-			),
-		);
-		// check width if device not found
-		if (!mobile && typeof window !== 'undefined') {
-			mobile = window.innerWidth <= deviceSize.mobileL;
-		}
-		setMobile(mobile);
-	}, []);
-
-	return { isMobile };
-}
-
 export function checkNetwork(networkId: number) {
 	const isXdai = networkId === xDaiChainId;
 	return networkId === appNetworkId || isXdai;
